requirements.txt
ConfigArgParse==1.7
diffusers==0.29.2
einops==0.8.1
einops_exts==0.0.4
GitPython==3.1.43
GitPython==3.1.44
gym==0.26.2
hydra-core==1.3.2
lightning_lite==1.8.6
matplotlib==3.7.5
MulticoreTSNE==0.1
MulticoreTSNE==0.1
numpy==1.23.0
numpy_quaternion==2023.0.4
omegaconf==2.1.2
opencv_python==4.6.0.66
opencv_python_headless==4.5.5.64
Pillow==11.1.0
plotly==5.22.0
pybullet==3.2.6
pyhash==0.9.3
pytorch_lightning==1.8.6
pyttsx3==2.98
scikit_learn==1.3.2
scipy==1.15.1
sentence_transformers==3.0.1
setuptools==57.5.0
tacto==0.0.3
termcolor==2.5.0
torch==2.3.1+cu118
torchvision==0.18.1+cu118
tqdm==4.66.4
transformers==4.37.2
wandb==0.19.6
